foodies-server - Bean Double to Decimal - wip

+++ src/main/java/bean/specificimplementation/ExtrasSpecificBeanImplementation.java
@@ -35,7 +35,7 @@ public class ExtrasSpecificBeanImplementation extends TableGenericBeanImplementa
             ShortName = "Precio",
             LongName = "Precio",
             Description = "Precio del plato",
-            Type = EnumHelper.FieldType.Double,
+            Type = EnumHelper.FieldType.Decimal,
             IsRequired = true,
             RegexPattern = "[0-9]{5,5}",
             RegexHelp = "5 dígitos",
